pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "okapi"
version = "0.1.0"
description = "Reinforcement Learning with JAX - PPO implementations for MuJoCo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "distrax>=0.1.7",
    "flax>=0.12.0",
    "gymnasium[box2d,mujoco]>=1.2.1",
    "hydra-core>=1.3.2",
    "jax[cuda]>=0.7.2",
    "matplotlib>=3.10.8",
    "mujoco>=3.3.7",
    "optax>=0.2.6",
    "orbax-checkpoint>=0.11.25",
    "playground>=0.0.5",
    "ruff>=0.14.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
    "warp-lang>=1.9.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/okapi"]
